+++ Lib/ToolsDir/jasmine-boot330.ts
@@ -189,6 +189,8 @@ declare var jasmineRequire: any;
         return result + arr_obj.join(", ") + "\n" + repeatString(_indent, stack_length - 1) + "}";
     }
 
+    let testId = window.location.hash;
+
     function realLog(message: string) {
         let stack: string;
         let err = <any>new Error();
@@ -200,7 +202,7 @@ declare var jasmineRequire: any;
                 stack = err.stack || err.stacktrace;
             }
         }
-        bbTest("consoleLog", { message, stack });
+        bbTest("consoleLog" + testId, { message, stack });
     }
 
     var bbTest = (<any>window.parent).bbTest;
@@ -218,7 +220,7 @@ declare var jasmineRequire: any;
             specFilter: specFilterFnc
         };
         onerror = ((msg: string, _url: string, _lineNo: number, _columnNo: number, error: Error) => {
-            bbTest("onerror", { message: msg, stack: error.stack });
+            bbTest("onerror" + testId, { message: msg, stack: error.stack });
         }) as any;
         env.configure(config);
         var perfnow: () => number;
@@ -240,11 +242,11 @@ declare var jasmineRequire: any;
         let totalStart = 0;
         env.addReporter({
             jasmineStarted: (suiteInfo: { totalSpecsDefined: number }) => {
-                bbTest("wholeStart", suiteInfo.totalSpecsDefined);
+                bbTest("wholeStart" + testId, suiteInfo.totalSpecsDefined);
                 totalStart = perfnow();
             },
             jasmineDone: () => {
-                bbTest("wholeDone", perfnow() - totalStart);
+                bbTest("wholeDone" + testId, perfnow() - totalStart);
                 var cov = (window as any).__c0v as Uint32Array;
                 if (cov != undefined) {
                     let pos = 0;
@@ -255,26 +257,26 @@ declare var jasmineRequire: any;
                             let len = maxlen - 1;
                             while (cov[pos + len] === 0) len--;
                             len++;
-                            bbTest("coverageReportPart", {
+                            bbTest("coverageReportPart" + testId, {
                                 start: pos,
                                 data: Array.prototype.slice.call(cov.slice(pos, pos + len))
                             });
                             pos += maxlen;
                             setTimeout(sendPart, 10);
                         } else {
-                            bbTest("coverageReportFinished", { length: cov.length });
+                            bbTest("coverageReportFinished" + testId, { length: cov.length });
                         }
                     };
-                    bbTest("coverageReportStarted", { length: cov.length });
+                    bbTest("coverageReportStarted" + testId, { length: cov.length });
                     setTimeout(sendPart, 10);
                 }
             },
             suiteStarted: (result: { description: string; fullName: string }) => {
-                bbTest("suiteStart", result.description);
+                bbTest("suiteStart" + testId, result.description);
                 stack.push(perfnow());
             },
             specStarted: (result: { description: string; stack: string }) => {
-                bbTest("testStart", { name: result.description, stack: result.stack });
+                bbTest("testStart" + testId, { name: result.description, stack: result.stack });
                 specStart = perfnow();
             },
             specDone: (result: {
@@ -283,7 +285,7 @@ declare var jasmineRequire: any;
                 failedExpectations: { message: string; stack: string }[];
             }) => {
                 let duration = perfnow() - specStart;
-                bbTest("testDone", {
+                bbTest("testDone" + testId, {
                     name: result.description,
                     duration,
                     status: result.status,
@@ -296,7 +298,7 @@ declare var jasmineRequire: any;
                 failedExpectations: { message: string; stack: string }[];
             }) => {
                 let duration = perfnow() - stack.pop()!;
-                bbTest("suiteDone", {
+                bbTest("suiteDone" + testId, {
                     name: result.description,
                     duration,
                     status: result.status,